# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hex Microservices API
  description: |
    Comprehensive API documentation for the Hex Microservices architecture.
    
    ## Architecture Overview
    - **API Gateway** (port 3000): Authentication and routing
    - **Hex Data Service** (port 3001): TCP data logging and management
    - **Notification Service** (port 3002): Push notifications and device management
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain tokens via `/api/auth/login`.
    
    ## Base URLs
    - API Gateway: `http://localhost:3000`
    - Hex Data Service: `http://localhost:3001` 
    - Notification Service: `http://localhost:3002`
  version: 1.0.0
  contact:
    name: Hex Microservices API
    email: developer@hexmicroservices.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: API Gateway (main entry point)
  - url: http://localhost:3001
    description: Hex Data Service
  - url: http://localhost:3002
    description: Notification Service
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: string
          description: Error message (present when success is false)
        message:
          type: string
          description: Additional information about the response
      required:
        - success
    HexLog:
      type: object
      properties:
        id:
          type: string
          description: Unique log entry identifier
        timestamp:
          type: string
          format: date-time
          description: When the log entry was created
        clientIP:
          type: string
          description: IP address of the client that sent the data
        dataLength:
          type: integer
          description: Length of the received data in bytes
        hexData:
          type: string
          description: Hexadecimal representation of the received data
        asciiData:
          type: string
          description: ASCII representation of the received data
        fileName:
          type: string
          description: File where this log entry is stored
      required:
        - id
        - timestamp
        - clientIP
        - dataLength
        - hexData
        - asciiData
        - fileName
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        username:
          type: string
          description: Username
        role:
          type: string
          enum: [user, admin]
          description: User role
        createdAt:
          type: string
          format: date-time
          description: When the user was created
      required:
        - id
        - username
        - role
        - createdAt
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: User's username
        password:
          type: string
          description: User's password
      required:
        - username
        - password
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          description: Desired username
        password:
          type: string
          description: User's password
        role:
          type: string
          enum: [user, admin]
          default: user
          description: User role
      required:
        - username
        - password
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "admin"
              password: "admin123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            description: JWT token
                          user:
                            $ref: '#/components/schemas/User'
              example:
                success: true
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    id: "admin-001"
                    username: "admin"
                    role: "admin"
                message: "Login successful"
        '400':
          description: Missing username or password
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: "newuser"
              password: "securepassword"
              role: "user"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                          user:
                            $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or missing fields
        '409':
          description: Username already exists
  /api/auth/verify:
    get:
      tags:
        - Authentication
      summary: Verify JWT token
      description: Verify if the provided JWT token is valid
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          userId:
                            type: string
                          role:
                            type: string
                          exp:
                            type: integer
                          iat:
                            type: integer
        '401':
          description: Invalid or expired token
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Get information about the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
        '404':
          description: User not found
  /api/auth/users:
    get:
      tags:
        - Authentication
      summary: List all users (Admin only)
      description: Get a list of all registered users (requires admin role)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Users list retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          count:
                            type: integer
        '401':
          description: Authentication required
        '403':
          description: Admin access required

  # Hex Data Service Endpoints
  /api/logs:
    get:
      tags:
        - Hex Logs
      summary: Get logs with pagination and filtering
      description: Retrieve hex logs with optional filtering by client IP and date range
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Number of logs per page
        - name: clientIP
          in: query
          schema:
            type: string
          description: Filter by client IP address
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter logs from this date (YYYY-MM-DD)
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter logs until this date (YYYY-MM-DD)
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order by timestamp
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          logs:
                            type: array
                            items:
                              $ref: '#/components/schemas/HexLog'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '500':
          description: Server error
  /api/logs/client/{clientIP}:
    get:
      tags:
        - Hex Logs
      summary: Get logs for specific client
      description: Retrieve all logs from a specific client IP address
      parameters:
        - name: clientIP
          in: path
          required: true
          schema:
            type: string
          description: Client IP address
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Client logs retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          logs:
                            type: array
                            items:
                              $ref: '#/components/schemas/HexLog'
                          total:
                            type: integer
                          clientIP:
                            type: string
...
  /api/logs/cleanup:
    delete:
      tags:
        - Hex Logs
      summary: Delete old log files
      description: Delete log files older than specified number of days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
          description: Delete files older than this many days
      responses:
        '200':
          description: Cleanup completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: integer
                            description: Number of files deleted
                          size:
                            type: string
                            description: Total size of deleted files

  # Notification Service Endpoints
...